+++ Faker/Config.cs
@@ -1,11 +1,13 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace RimuTec.Faker
+﻿namespace RimuTec.Faker
 {
+   /// <summary>
+   /// Configuration for RimuTec.Faker
+   /// </summary>
    public static class Config
    {
+      /// <summary>
+      /// Gets/sets the locale or customization to be used by generators.
+      /// </summary>
       public static string Locale { get; set; } = "en"; // "en" is default
    }
 }